DR0053A-149 PR improvements
