+++ Pipfile
@@ -5,7 +5,7 @@ name = "pypi"
 
 [packages]
 pyside6 = "==6.6.0"
-ingenialink = "==7.1.0"
+ingenialink = "==7.1.1"
 ingeniamotion = "==0.7.0"
 ingenialogger = "==0.3.0"
